Add react and webpack
This is part of a gradual transition to webpack and react in an effort
to add components. While webpack and react have been tested as working,
nothing has currently been switched to them.

Gulp is still used but may be replaced with webpack + npm scripts

The directory structure has been changed to minimize building. It may
change more when components are added.