Moved header into app
The header is now part of the react app which means lodash is in and
underscore is out and that there is no flicker of unstyled header.

Javascript building is now done with webpack instead of gulp.

The triangles have not yet been converted to react, but they do draw
more efficiently/less frequently now.