Add timeline component
The timeline on the resume is now built from an array of data objects.

It can now automatically determine which timeline points to hide and
what number to display in the more/less button.

For now this component will be assumed to be specific to the resume,
despite being made somewhat generically, and could eventually be a
general component.