Separate development and production building
Why was this not done before? Before webpack I was able to
convert/compress for production in development without a performance
hit.

However, it is much better to separate the two different scenarios for
building and should work nicely as I add new features specific to either
scenario.

React correctly reports that it is in development mode or production
mode depending on which build was run.