Add npm script for webpack watch
Unfortunately there is no way that I am aware of to run webpack watch
mode without it starting with a build.

Be careful not to run a build while the server is running (after
docker-compose up). The build will run the webpack clean plugin which
will clear out the build directory which will break the server because
it thinks everything is gone. Your options are to `npm run watch` then
in a separate bash `docker-compose up` once the former is done or to run
`npm run build:dev && docker-compose up` for a one-time build startup.

This is a minor inconvenience that can be ignored for now because the
benefit of watch based mini builds outweight this problem.