Convert uploader to react
I plan to break out the file uploader into a site-wide component because
it should be despite there not being anything else using it.

Users that click the drag n drop area will now be able to select more
than one file to upload at a time. This feature was not present before.

Also added an uploading state. There is room for improvement here such
as saying how many are uploaded versus how many to upload, and checking
for if there are no more pending uploads before removing uploading
state.