Upgrade html-webpack-plugin
On older versions of html-webpack-plugin, it would return broken html if
not injecting its own script. An unmaintained plugin, that wasn't
working with webpack4, was needed to add defer if it was injecting its
own script. My temporary solution was to let it inject and manually add
defer until it was fixed and now it is.