Add next-gen image format
Production builds will now output webp for all png and jpg. pngs are 
lossless and jpgs are of quality 65.

A picture component has been added to use the picture html element to 
provide a fallback from webp to png. I didn't add jpg to this component 
yet and only used it in one place. The component will load pngs for the 
develop environment where webps are not output.