Change commonjs to esm
ecmascript modules have lost the experimental flag in node 13 and are gaining library support.

I have added the latest eslint to support cjs. jest has been updated to 24, and has had cjs supported merged in but not released. In the meantime the jest config has been added to the package.json