Use yarn instead of npm lockfile
As much as I'd like to not have yarn because it is one more dependency, it is just better. Yarn is faster, works better when locally testing one project in another, and is much simpler to use. The breaking point for me today was when npm gave me an ugly error because I tried to remove a package when offline. Yarn happily removed the package while offline without an error.