Add touch events to growth graphs
the css for hover wont trigger on touch, so a class is toggled via js instead.

touchmove, unlike mousemove, can happen outside the bounds of the element so a check to see if we are moving over the element is necessary

checked via simulated touch events in the dev tools.